feat: package cowin and notify; introduce go mod
