+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/4j4y/cowing
+
+go 1.16
+
+require github.com/gen2brain/beeep v0.0.0-20200526185328-e9c15c258e28
